Refactor: MerkleProofAccess
